+++ CHANGELOG.md
@@ -26,6 +26,7 @@ Changes are grouped as follows
 
 ### Added
 - `limit` parameter on sequence data retrieval.
+- Support for relationships exposed through experimental client.
 - `end` parameter of sequence.data retrieval and range delete accepts -1 to indicate last index of sequence.
 
 ### Fixed